Implement queue removal functions
Added q_remove_head function to remove the head of the queue and
return its element_t structure.
Added q_remove_tail function to remove the tail of the queue and
return its element_t structure.

These functions both return NULL if queue is NULL or is empty.

If the output buffer is smaller than the removed string, the value
is truncated, and a null terminator is ensured at the end of the buffer.

Change-Id: I4678c8fa89e4faacc58bdbe76c9b14c3e254e5e1